Fix: form component 삭제
